[refactor] cast column data

+++ src/main/java/ecnu/db/schema/Column.java
@@ -164,7 +164,7 @@ public class Column {
         switch (columnType) {
             case DATE, DATETIME -> {
                 for (int i = 0; i < columnData.length; i++) {
-                    ret[i] = columnData[i] + min;
+                    ret[i] = (double) columnData[i] + min;
                 }
             }
             case DECIMAL -> {
